reactlibapp/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
coreapi==2.3.1
coreschema==0.0.4
coverage==4.4.1
coveralls==1.1
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.3
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-envie==0.0.9
django-filter==1.0.4
django-nose==1.4.4
django-postgrespool==0.3.1
django-rest-swagger==2.1.2
django-toolbelt==0.0.1
django-webpack-loader==0.5.0
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
docopt==0.6.2
google-api-python-client==1.6.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
nose==1.3.7
oauth2client==4.1.2
openapi-codec==1.3.2
psycopg2==2.7.3
pyasn1==0.3.2
pyasn1-modules==0.0.11
PyJWT==1.5.2
pystache==0.5.4
python-env==1.0.0
pytz==2017.2
PyYAML==3.12
requests==2.18.3
rsa==3.4.2
simplejson==3.11.1
six==1.10.0
SQLAlchemy==1.1.12
sqlparse==0.2.3
static==1.1.1
static3==0.7.0
uritemplate==3.0.0
urllib3==1.22
